requirements.txt
requests==2.28.2
python-decouple==3.7
shapely==2.0.0
